requirements.txt
Django==5.2
psycopg[binary]==3.2.13
python-decouple==3.8
django-taggit==6.1.0
Markdown==3.7
djangorestframework==3.16.0
django-redis==5.4.0
# Pillow==11.2.1

# Dev tools
pre-commit==4.2.0
black==25.1.0
isort==6.0.1
flake8==7.2.0
pytest==8.3.5
pytest-django==4.10.0
coverage==7.8.0
